+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 - rpc: serialize to `Vec<u8>` instead of `serde_json::Value` ([#38](https://github.com/lamports-dev/alpamayo/pull/38))
+- rpc: remove get_block concurrency ([#39](https://github.com/lamports-dev/alpamayo/pull/39))
 
 ### Breaking